Rename to match component
more check run list item rename
