+++ app/src/ui/check-runs/ci-check-run-list-item.tsx
@@ -3,7 +3,7 @@ import { IRefCheck } from '../../lib/stores/commit-status-store'
 import { Octicon } from '../octicons'
 import { getClassNameForCheck, getSymbolForCheck } from '../branches/ci-status'
 import classNames from 'classnames'
-import { CICheckRunLogs } from './ci-check-run-item-logs'
+import { CICheckRunLogs } from './ci-check-run-logs'
 
 interface ICICheckRunListItemProps {
   /** The check run to display **/
